Rename tests file
