+++ packages/core/test/index.ts
@@ -78,7 +78,7 @@ import "./tag/compoundTagTests";
 import "./tag/tagTests";
 import "./text/textTests";
 import "./toast/overlayToasterTests";
-import "./toast/toast2Tests";
+import "./toast/toastTests";
 import "./tooltip/tooltipTests";
 import "./tree/treeTests";
